Fixed issues with vectors of BackupPrivateCopy
If a vector capacity is exceeded, its items are moved to the new
location.
By default, a move is performed using a move constructor and
later destroying the old object. BackupPrivateCopies were
being moved using default move constructor, which just copies its
contents. Later on, the destructor was releasing the memory, so
when a vector was being reallocated, its BackupPrivateCopy items
were deallocating the memory twice (of course this is illegal).

Vector reallocation is not necessary if there is enough space for
all items. The problem was that the capacity reservation was not
placed at the proper place (MemController constructor).

Fixed this problem declaring a move constructor for BackupPrivateCopy
and by placing the vector::reserve calls in the proper place.