Moving recovery logic from SMPDD to WorkDescriptor
Most of recovery logic that was previously done in SMPDD::execute
is compatible with nearly all the devices (checking if a task has
been unsuccessfully executed several times, if its parent has
been invalidated, etc.). Therefore, that logic is moved to the
WorkDescriptor, which makes more sense.
nanos::error::ExecutionFailure is now in charge of invalidating
the Workdescriptor, and checking if it has run out of re-execution
trials (in such case it just invalidates its ancestors and gives
up retrying).