Change MemoryPage::remap and add page lock/unlock
Remap was using unmap and mmap to an specific memory page. This is
not completely wrong, but is sensible to a data race: if another
thread allocates memory, it might reserve the virtual address mapped
by that specific page.

Using mmap with MAP_FIXED flag, which discards the previous mapping
pointed by the 'hint' address. In addition, it fails when mapping to
the provided address is not possible.

Adding lock()/unlock() methods (not used currently) that allow the
page to be pinned/unpinned.