Added ensure to check if FPGA task copies are valid
If a FPGA task has copies not registered in the internal FPGA allocator the runtime sends a NULL xdma handler to the xdma library. This happens when the task copies are portions of a malloc region instead of xdma_malloc region. The ensure makes the application fall in the debug mode when the described situation happens.